Search Refactor
Refactored the search so it functions as an independent class.

An instance of the class is utilized within UCI Handling. This was done in preparation for datagen / multi-thread setup, as well as getting tables out of global memory.

Elo: 6.89 +/- 3.89, nElo: 10.71 +/- 6.05
LOS: 99.97 %, DrawRatio: 50.54 %, PairsRatio: 1.10
Games: 12656, Wins: 4737, Losses: 4486, Draws: 3433, Points: 6453.5 (50.99 %)
Ptnml(0-2): [303, 1186, 3198, 1239, 402]
LLR: 2.90 (-2.25, 2.89) [0.00, 3.00]